pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.2.0",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "notebook>=7.4.3",
    "python-dotenv>=1.1.0",
]
